reorganize codebase to support adding UI
