+++ cloudformation/main.yml
@@ -86,7 +86,7 @@ Resources:
       Runtime: nodejs20.x
       Description: !Sub "${ApplicationFriendlyName} API Lambda"
       FunctionName: !Sub ${ApplicationPrefix}-lambda
-      Handler: lambda.handler
+      Handler: api.lambda.handler
       MemorySize: 512
       Role: !GetAtt AppSecurityRoles.Outputs.MainFunctionRoleArn
       Timeout: 60
