+++ src/api/routes/protected.ts
