+++ tests/unit/auth.test.ts
@@ -4,7 +4,7 @@ import {
   SecretsManagerClient,
 } from "@aws-sdk/client-secrets-manager";
 import { mockClient } from "aws-sdk-client-mock";
-import init from "../../src/index.js";
+import init from "../../src/api/index.js";
 import {
   secretJson,
   secretObject,
