+++ tests/unit/organizations.test.ts
@@ -1,5 +1,5 @@
 import { afterAll, expect, test } from "vitest";
-import init from "../../src/index.js";
+import init from "../../src/api/index.js";
 
 const app = await init();
 test("Test getting the list of organizations succeeds", async () => {
